fix: cherry-pick upstream Darling Fedora 44 build fixes
Five small portability fixes from darlinghq/darling master (not yet in this fork's
base f39a29489), all superproject framework sources, no build-system impact:
- e3fe42888 DiskArbitration DADisk.c (CFStringGetCStringPtr)
- 3f277ba53 libaks (aks_save_bag size param int*)
- 9f485c91d ImageIO CGImageSource.m (ObjC casts)
- ddd118d9b SecurityFoundation header (decrypt: param type)
- fc5c0666b OpenDirectory ODSession.m
Help modern (Fedora 44 era) toolchains build these frameworks.

+++ src/libaks/libaks.c
@@ -66,7 +66,7 @@ kern_return_t aks_create_bag(uint8_t* secret, int secret_size, int bag_type, key
 	return KERN_FAILURE;
 };
 
-kern_return_t aks_save_bag(keybag_handle_t handle, void** bytes, size_t* size) {
+kern_return_t aks_save_bag(keybag_handle_t handle, void** bytes, int* size) {
     printf("STUB %s\n", __PRETTY_FUNCTION__);
 	return KERN_FAILURE;
 };
